Master

+++ app/src/main/java/com/shashankbhat/splitbill/ui/user_list/UserListFragment.kt
@@ -4,18 +4,18 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.fragment.app.viewModels
 import androidx.navigation.NavController
 import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.LinearLayoutManager
+import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import com.shahankbhat.recyclergenericadapter.RecyclerGenericAdapter
 import com.shahankbhat.recyclergenericadapter.util.CallBackModel
 import com.shahankbhat.recyclergenericadapter.util.DataBinds
 import com.shahankbhat.recyclergenericadapter.util.MoreDataBindings
 import com.shashankbhat.splitbill.BR
 import com.shashankbhat.splitbill.R
-import com.shashankbhat.splitbill.base.TitleFragment
+import com.shashankbhat.splitbill.base.BaseFragment
 import com.shashankbhat.splitbill.database.local.dto.group_list.GroupListDto
 import com.shashankbhat.splitbill.database.local.entity.User
 import com.shashankbhat.splitbill.databinding.AdapterGroupUserBinding
@@ -25,7 +25,7 @@ import com.shashankbhat.splitbill.viewmodels.UserListViewModel
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
-class UserListFragment : TitleFragment() {
+class UserListFragment : BaseFragment() {
     private lateinit var binding: FragmentUserListBinding
     private lateinit var navController: NavController
     private lateinit var groupListDto: GroupListDto
@@ -46,7 +46,6 @@ class UserListFragment : TitleFragment() {
         return binding.root
     }
 
-    @OptIn(ExperimentalComposeUiApi::class)
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
@@ -56,9 +55,17 @@ class UserListFragment : TitleFragment() {
         viewModel.getAllUsersByGroupId(groupListDto.group?.id ?: -1)
         viewModel.getAllBill(groupListDto.group?.id ?: -1)
 
-        binding.fab.setOnClickListener {
+        uiFabClickListener()
+        uiRecyclerViewInit()
 
-            val addMember = AddGroupMemberFragment(viewModel, groupListDto)
+        networkUserListResponse()
+    }
+
+    private fun uiFabClickListener(){
+        binding.fab.setOnClickListener {
+            val addMember = AddGroupMembersFragment.newInstance {
+                viewModel.addPeople(User(it, groupListDto.group?.id ?: -1))
+            }
             context?.findActivity()?.supportFragmentManager?.let {
                 addMember.show(
                     it,
@@ -67,12 +74,16 @@ class UserListFragment : TitleFragment() {
             }
 
         }
+    }
 
+    private fun uiRecyclerViewInit(){
         adapter = RecyclerGenericAdapter.Builder<AdapterGroupUserBinding, User>(R.layout.adapter_group_user, BR.model)
             .setClickCallbacks(arrayListOf<CallBackModel<AdapterGroupUserBinding, User>>().apply {
                 add(CallBackModel(R.id.iv_link){ model, _, _ ->
-                    val dialog = LinkGroupMemberFragment(viewModel, model)
-                    requireContext().findActivity()?.supportFragmentManager?.let {
+                    val dialog = LinkGroupMemberDialogFragment.newInstance {
+                        viewModel.linkPeople(model, it)
+                    }
+                    context?.findActivity()?.supportFragmentManager?.let {
                         dialog.show(
                             it,
                             dialog.tag
@@ -80,7 +91,13 @@ class UserListFragment : TitleFragment() {
                     }
                 })
                 add(CallBackModel(R.id.iv_delete){ model, _, _ ->
-                    viewModel.deleteUser(model)
+                    val deleteDialog = MaterialAlertDialogBuilder(requireContext())
+                        .setTitle("Delete")
+                        .setMessage("Are you sure you want to delete ${model.name}")
+                        .setPositiveButton("Delete") { _, _ ->
+                            viewModel.deleteUser(model)
+                        }.setNegativeButton("Cancel", null)
+                    deleteDialog.show()
                 })
             })
             .setMoreDataBinds(DataBinds().apply {
@@ -90,7 +107,9 @@ class UserListFragment : TitleFragment() {
 
         binding.rvNearbyUsers.layoutManager = LinearLayoutManager(requireContext())
         binding.rvNearbyUsers.adapter = adapter
+    }
 
+    private fun networkUserListResponse(){
         viewModel.userListState.observe(viewLifecycleOwner) {
             if(it.isSuccess()){
                 if(adapter.getItemList().size == it.data?.size){
@@ -107,8 +126,6 @@ class UserListFragment : TitleFragment() {
                     adapter.replaceList(ArrayList(it.data ?: emptyList()))
             }
         }
-
-
     }
 
     companion object {
